+++ R/authentication.R
@@ -168,6 +168,8 @@ BasicAuth <- R6Class(
 #' \itemize{
 #'   \item{authorization_code}
 #'   \item{authorization_code+pkce}
+#'   \item{client_credentials}
+#'   \item{urn:ietf:params:oauth:grant-type:device_code}
 #'   \item{urn:ietf:params:oauth:grant-type:device_code+pkce}
 #' }
 #' 
@@ -252,9 +254,9 @@ AbstractOIDCAuthentication <- R6Class(
         private$scopes = list("openid")
       } else {
         private$scopes = provider$scopes
-        
+
         #TODO remove later, this is used for automatic reconnect
-        if (!"offline_access" %in% private$scopes) {
+        if (private$grant_type != "client_credentials" && !"offline_access" %in% private$scopes) {
           private$scopes = c(private$scopes, "offline_access")
         }
       }
@@ -267,8 +269,8 @@ AbstractOIDCAuthentication <- R6Class(
       
       private$client_id = config$client_id
       
-      if (private$grant_type == "authorization_code") {
-        # in this case we need a client_id and secrect, which is basically the old OIDC Auth Code implementation
+      if (private$grant_type == "authorization_code" || private$grant_type == "client_credentials") {
+        # in this case we need a client_id and secrect
         if (!all(c("client_id","secret") %in% names(config))) {
           stop("'client_id' and 'secret' are not present in the configuration.")
         }
@@ -280,7 +282,6 @@ AbstractOIDCAuthentication <- R6Class(
           secret = config$secret
         )
       } else {
-      
         private$oauth_client = oauth_client(
           id = private$client_id,
           token_url = private$endpoints$token_endpoint,
@@ -446,16 +447,9 @@ OIDCDeviceCodeFlow <- R6Class(
   public = list(
     # functions ####
     login = function() {
-      
-      client <- oauth_client(
-        id = private$client_id,
-        token_url = private$endpoints$token_endpoint,
-        name = "openeo-r-oidc-auth"
-      )
-
       private$auth = rlang::with_interactive(
                       oauth_flow_device(
-                        client = client,
+                        client = private$oauth_client,
                         auth_url = private$endpoints$device_authorization_endpoint,
                         scope = paste0(private$scopes, collapse = " ")
                       ),
@@ -488,16 +482,9 @@ OIDCDeviceCodeFlowPkce <- R6Class(
   public = list(
     # functions ####
     login = function() {
-      
-      client <- oauth_client(
-        id = private$client_id,
-        token_url = private$endpoints$token_endpoint,
-        name = "openeo-r-oidc-auth"
-      )
-
       private$auth = rlang::with_interactive(
                       oauth_flow_device(
-                        client = client,
+                        client = private$oauth_client,
                         auth_url = private$endpoints$device_authorization_endpoint,
                         scope = paste0(private$scopes, collapse = " "),
                         pkce = TRUE
@@ -533,16 +520,9 @@ OIDCAuthCodeFlowPKCE <- R6Class(
     # attributes ####
     # functions ####
     login = function() {
-
-      client <- oauth_client(
-        id = private$client_id,
-        token_url = private$endpoints$token_endpoint,
-        name = "openeo-r-oidc-auth"
-      )
-
       private$auth = rlang::with_interactive(
                       oauth_flow_auth_code(
-                        client = client,
+                        client = private$oauth_client,
                         auth_url = private$endpoints$authorization_endpoint,
                         scope = paste0(private$scopes, collapse = " "),
                         pkce = TRUE,
@@ -580,15 +560,9 @@ OIDCAuthCodeFlow <- R6Class(
     
     # functions ####
     login = function() {
-      client <- oauth_client(
-        id = private$client_id,
-        token_url = private$endpoints$token_endpoint,
-        name = "openeo-r-oidc-auth"
-      )
-
       private$auth = rlang::with_interactive(
                       oauth_flow_auth_code(
-                        client = client,
+                        client = private$oauth_client,
                         auth_url = private$endpoints$authorization_endpoint,
                         scope = paste0(private$scopes, collapse = " "),
                         pkce = FALSE,
@@ -610,6 +584,37 @@ OIDCAuthCodeFlow <- R6Class(
       
       if (!any(c("authorization_code") %in% grant_types)) {
         stop("Authorization code flow is not supported by the authentication provider")
+      }
+      invisible(TRUE)
+    }
+  )
+)
+
+# [OIDCClientCredentialsFlow] ----
+OIDCClientCredentialsFlow <- R6Class(
+  "OIDCClientCredentialsFlow",
+  inherit = AbstractOIDCAuthentication,
+  # public ====
+  public = list(
+    # functions ####
+    login = function() {
+      private$auth = oauth_flow_client_credentials(
+                        client = private$oauth_client,
+                        scope = paste0(private$scopes, collapse = " ")
+                      )
+
+      invisible(self)
+    }
+  ),
+  # private ====
+  private = list(
+    # attributes ####
+    grant_type = "client_credentials", # not used internally by httr2, but maybe useful in openeo
+    
+    # functions ####
+    isGrantTypeSupported = function(grant_types) {
+      if (!"client_credentials" %in% grant_types) {
+        stop("Client Credentials flow is not supported by the authentication provider")
       }
       invisible(TRUE)
     }