+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Added
 
+* Support for the OpenID Connect Client Credentials flow
 * compute_result, create_job, update_job, create_service and update_service have an additional parameter to inject custom properties into the request (e.g., to set memory or CPU limits)
 
 # Version 1.3.1